+++ app/models/users/RegisteredUser.java
@@ -22,7 +22,7 @@ public class RegisteredUser extends User{
     private String town;
     private String postCode;
     private String creditCard;
-    private Double amount=0f;
+    private Float amount=0f;
 
     // RegisteredUser has one basket.
     // RegisteredUser is the owner (foreign key will be added to Basket
@@ -44,7 +44,7 @@ public class RegisteredUser extends User{
         this.creditCard="";
         this.amount=0f;
     }
-	public RegisteredUser(String email, String role, String name, String password, String street1, String street2, String town, String postCode, String creditCard, Double amount)
+	public RegisteredUser(String email, String role, String name, String password, String street1, String street2, String town, String postCode, String creditCard, Float amount)
 	{
 		super(email, role, name, password);
         this.street1 = street1;
@@ -86,13 +86,13 @@ public class RegisteredUser extends User{
     public void setPostCode(String postCode) {
         this.postCode = postCode;
     }
-    public Double getWallet(){
+    public Float getWallet(){
         return this.amount;
     }
-    public void setWallet(Double amount){
+    public void setWallet(Float amount){
         this.amount=amount;
     }
-    public void addWallet(Double amount){
+    public void addWallet(Float amount){
         this.amount+=amount;
     }
